+++ trinity/rpc/modules/evm.py
@@ -2,18 +2,18 @@ from eth_utils import (
     encode_hex,
 )
 
-from evm.rpc.format import (
-    format_params,
-)
-from evm.rpc.modules import (
-    RPCModule,
-)
-
-from evm.utils.fixture_tests import (
+from evm.tools.fixture_tests import (
     apply_fixture_block_to_chain,
     new_chain_from_fixture,
     normalize_block,
     normalize_blockchain_fixtures,
+)
+
+from trinity.rpc.format import (
+    format_params,
+)
+from trinity.rpc.modules import (
+    RPCModule,
 )
 
 
@@ -22,7 +22,7 @@ class EVM(RPCModule):
     def resetToGenesisFixture(self, chain_info):
         '''
         This method is a special case. It returns a new chain object
-        which is then replaced inside :class:`~evm.rpc.main.RPCServer`
+        which is then replaced inside :class:`~trinity.rpc.main.RPCServer`
         for all future calls.
         '''
         return new_chain_from_fixture(chain_info)
@@ -31,7 +31,7 @@ class EVM(RPCModule):
     def applyBlockFixture(self, block_info):
         '''
         This method is a special case. It returns a new chain object
-        which is then replaced inside :class:`~evm.rpc.main.RPCServer`
+        which is then replaced inside :class:`~trinity.rpc.main.RPCServer`
         for all future calls.
         '''
         _, _, rlp_encoded = apply_fixture_block_to_chain(block_info, self._chain)
